move file from qkc to cluster/master/
